+++ js/util.js
@@ -1,0 +1,20 @@
+// eslint-disable-next-line no-unused-vars
+function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+function maxLenStr(str, maxLen)
+{
+  return str.length <= maxLen;
+}
+
+maxLenStr('', 1);
+
+function getRandomArrayElement(ARR) {
+  return ARR[getRandomPositiveInteger(0, ARR.length-1)];
+}
+
+export {getRandomArrayElement, getRandomPositiveInteger, maxLenStr};
